backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.4
httpx==0.28.1
python-multipart==0.0.20
